requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chirpstack-api==4.6.0
google-api-core==2.15.0
google-auth==2.25.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
idna==3.6
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
requests==2.31.0
rsa==4.9
urllib3==2.1.0
